% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_cir.R
\name{cre_cir}
\alias{cre_cir}
\title{Hi}
\usage{
cre_cir(T_num = 5, S_num = 25, S_len = 2, angle = 360)
}
\arguments{
\item{T_num}{hi}

\item{S_num}{hi}

\item{S_len}{hi}

\item{angle}{hi}
}
\value{

}
\description{
Hi
}
\examples{
cre_cir()
}
