% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_cran.R
\name{cre_cran}
\alias{cre_cran}
\title{Hi}
\usage{
cre_cran(T_len = 2, T_wid = 2, S_num = 18, angle = 360, seed)
}
\arguments{
\item{T_len}{Hi}

\item{S_num}{HI}
}
\value{

}
\description{
Hi
}
\examples{
cre_cran()
}
