% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cre_tran.R
\name{cre_tran}
\alias{cre_tran}
\title{Title Hi}
\usage{
cre_tran(T_len = 60, T_wid = 7, S_num = 18, seed)
}
\arguments{
\item{T_len}{Hi}

\item{T_wid}{Hi}

\item{S_num}{Hi}
}
\value{

}
\description{
Title Hi
}
\examples{
cre_tran()
}
