% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_twidth.R
\name{find_twidth}
\alias{find_twidth}
\title{Find minimum transect width}
\usage{
find_twidth(
  Shape = "rec",
  T_Len = 60,
  T_num = 2,
  S_num = 18,
  S_Wid = 2,
  S_Len = 2,
  S_Bor = 5
)
}
\arguments{
\item{Shape}{Shape of the site}

\item{T_Len}{Length of transect (in m)}

\item{T_num}{Number of transects in a group}

\item{S_num}{Number of sites}

\item{S_Wid}{Width of each site (in m)}

\item{S_Len}{Length of each site (in m)}

\item{S_Bor}{Border of each site (in m)}
}
\value{

}
\description{
Find minimum transect width
}
\examples{
find_twidth()
}
