% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_geo.R
\name{graph_geo}
\alias{graph_geo}
\title{Plots lon, lat coordinate in ggplot}
\usage{
graph_geo(df, geo = F)
}
\arguments{
\item{df}{dataframe with Type, S_lon and S_lat}

\item{geo}{draws Natural Earth map}
}
\value{

}
\description{
Plots lon, lat coordinate in ggplot
}
\examples{
graph_geo(trs_geo(cre_tran(), SP = c(0,0), BRNG = 0))
}
