% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graph_loc.R
\name{graph_loc}
\alias{graph_loc}
\title{Plots x, y coordinate in ggplot}
\usage{
graph_loc(df)
}
\arguments{
\item{df}{dataframe with Type, S_x and S_y}
}
\value{

}
\description{
Plots x, y coordinate in ggplot
}
\examples{
graph_loc(cre_tran())
}
