% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs_geo.R
\name{trs_geo}
\alias{trs_geo}
\title{Find geoloction of transect}
\usage{
trs_geo(df, SP, BRNG, EP)
}
\arguments{
\item{df}{Dataframe, that consist of results from find_twidth or find_theight}

\item{SP}{Starting Point, in the format of c(Lat, Lon)}

\item{BRNG}{Bearing of SP to EP, only accepts numeric values}

\item{EP}{Ending Point (expected), in the format of c(Lat, Lon). Ignored if BRNG is not NA.}
}
\value{

}
\description{
Find geoloction of transect
}
\examples{
trs_geo(df, SP = c(10, 10))
}
