% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs_join.R
\name{trs_join}
\alias{trs_join}
\title{HI}
\usage{
trs_join(obj1, obj2, shiftx = 0, shifty = 0)
}
\arguments{
\item{obj1}{hi}

\item{obj2}{hi}

\item{shiftx}{hi}

\item{shifty}{hi}
}
\value{

}
\description{
HI
}
\examples{
trs_join(cre_twid(), cre_twid(), shiftx = 1, shifty = 1)
}
