% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trs_rot.R
\name{trs_rot}
\alias{trs_rot}
\title{Find X, Y coordinate after rotation}
\usage{
trs_rot(obj, rot)
}
\arguments{
\item{obj}{tranhelper list object}

\item{rot}{rotation angle (in deg)}
}
\value{

}
\description{
Rotates around 0,0 for tranhelper object
}
\examples{
trs_rot(cre_twid(), 90)
}
