% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic_findloc.R
\name{vic_findloc}
\alias{vic_findloc}
\title{Title}
\usage{
vic_findloc(p1, brng, dis, a = 6378137, f = 1/298.257223563)
}
\arguments{
\item{p1}{location in the format of c(lat, lon)}

\item{brng}{initial bearing}

\item{dis}{distance in m}
}
\value{

}
\description{
Title
}
\examples{
vic_finddis(c(10,10), 30, 18000)
}
