% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_obj.R
\name{write_obj}
\alias{write_obj}
\title{Exports transhelper calculated results as dataframe or csv}
\usage{
write_obj(obj, file)
}
\arguments{
\item{obj}{transhelper object}

\item{file}{file name}
}
\value{

}
\description{
Exports transhelper calculated results as dataframe or csv
}
\examples{
write_obj(find_twidth())
}
