% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vic_finddis.R, R/vic_finddis2.R
\name{vic_finddis}
\alias{vic_finddis}
\title{Finds distance between 2 points}
\usage{
vic_finddis(p1, p2, obj, a = 6378137, f = 1/298.257223563)

vic_finddis(p1, p2, obj, a = 6378137, f = 1/298.257223563)
}
\arguments{
\item{p1}{Location of point 1, in the format of c(lat, lon). Both in degrees}

\item{p2}{Location of point 2, in the format of c(lat, lon). Both in degrees}

\item{obj}{A tranhelper object, that contains all the x,y coordinates}

\item{a}{constant}

\item{f}{constant}
}
\value{



}
\description{
Via Vincenty's inverse formula
A direct translation of https://nathanrooy.github.io/posts/2016-12-18/vincenty-formula-with-python/ from python to R
Thanks ;D

Via Vincenty's inverse formula
A direct translation of https://nathanrooy.github.io/posts/2016-12-18/vincenty-formula-with-python/ from python to R
Thanks ;D
}
\examples{
vic_finddis((10, 10), (20, 20))

vic_finddis((10, 10), (20, 20))

}
